bbox_fusion/requirements.txt
ensemble-boxes
paho-mqtt
numpy
Pillow
PyYAML
requests
psutil
